+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 dart pub global activate fluttercraft
 
 # From binary
-# Download fluttercraft-v0.0.6-windows-x64.exe from releases
+# Download fluttercraft.exe from releases
 ```
 
 ## 0.0.5 (2025-12-15)